Heal'm App IOS (Protoype)
This is all the app development for the GPIP of 2023. The UI could be improved although keep it simplistic. The Login, SignUp, and Verification Email work with FireBase. The Bluetooth Functionality is working for only one microcontroller device, although there is the capability for multiple microcontroller devices to be connected. The emergency contacts are functional and do not need to be changed since it uses CoreData, and are based on user input. Further improvements that can be done are the settings page since the bottom section is not functional yet and automatic wound detection which calls 911 automatically is something that is to be added further. There is more functionality features to be added, but that is up to Prof. Dieff Vital for further instruction.